qemu: Move qemu/provider_lib -> provider
This is a leftover from the previous specification of
a test provider. In fact, the common provider lib should
be at top level, and the name of the lib is provider,
not provider_lib. Fix tests accordingly.
