+++ qemu/tests/migration_multi_host_ping_pong.py
@@ -4,7 +4,7 @@ from autotest.client.shared import error
 from autotest.client import utils
 from virttest import env_process, utils_test, remote, virt_vm, utils_misc
 from autotest.client.shared.syncdata import SyncData
-from provider_lib import cpuflags
+from provider import cpuflags
 
 
 @error.context_aware
